Extract table into own module
