+++ frontend/lib/src/components/core/Block/ElementNodeRenderer.tsx
@@ -95,7 +95,7 @@ import {
 } from "./utils"
 
 // Lazy-load elements.
-const ArrowTable = lazy(() => import("~lib/components/elements/ArrowTable"))
+const Table = lazy(() => import("~lib/components/elements/Table"))
 const ArrowVegaLiteChart = lazy(
   () => import("~lib/components/elements/ArrowVegaLiteChart")
 )
@@ -229,7 +229,7 @@ const RawElementNodeRenderer = (
           config={ElementContainerConfig.LARGE_ELEMENT}
           isStale={isStale}
         >
-          <ArrowTable
+          <Table
             element={tableProto}
             data={node.quiverElement}
             {...elementProps}
