+++ frontend/lib/src/components/elements/Table/styled-components.ts
