+++ lib/tests/streamlit/elements/table_test.py
@@ -258,3 +258,20 @@ class ArrowTest(DeltaGeneratorTestCase):
         assert element.width_config.pixel_width == 400
         assert element.height_config.WhichOneof("height_spec") == "pixel_height"
         assert element.height_config.pixel_height == 300
+
+
+class StTableAPITest(DeltaGeneratorTestCase):
+    """Test Public Streamlit Public APIs."""
+
+    def test_table(self):
+        """Test st.table."""
+        from streamlit.dataframe_util import convert_arrow_bytes_to_pandas_df
+
+        df = pd.DataFrame([[1, 2], [3, 4]], columns=["col1", "col2"])
+
+        st.table(df)
+
+        proto = self.get_delta_from_queue().new_element.table
+        pd.testing.assert_frame_equal(
+            convert_arrow_bytes_to_pandas_df(proto.arrow_data.data), df
+        )
